renamed the ProductCards folder to Products
