+++ src/components/ShoppingCart/index.jsx
@@ -1,7 +1,7 @@
 import styles from "./ShoppingCart.module.css";
 import { useLocation, useOutletContext } from "react-router-dom";
 import useScreenSizeCheck from "../../hooks/useScreenSizeCheck";
-import ProductsInCart from "../ProductCards/ProductsInCart";
+import ProductsInCart from "../Products/ProductsInCart";
 import SumTotal from "../SumTotal";
 import Button from "../Button";
