+++ src/routes/Home/index.jsx
@@ -1,7 +1,7 @@
 import { Helmet, HelmetProvider } from "react-helmet-async";
 import styles from "./Home.module.css";
 import BannerCommercial from "../../components/BannerCommercial";
-import ProductsToDisplay from "../../components/ProductCards/ProductLists";
+import ProductsToDisplay from "../../components/Products/ProductLists";
 import Categories from "../../components/Categories";
 
 export default function Home() {
