+++ src/routes/ProductSpecific/index.jsx
@@ -4,7 +4,7 @@ import { Helmet, HelmetProvider } from "react-helmet-async";
 import styles from "./ProductSpecific.module.css";
 import useFetch from "../../hooks/useFetch.jsx";
 import BreadCrumbs from "../../components/BreadCrumbs";
-import SingleProduct from "../../components/ProductCards/SingleProduct";
+import SingleProduct from "../../components/Products/SingleProduct/index.jsx";
 
 const url = import.meta.env.VITE_API_BASE_URL;
 export const ThisProductContext = createContext();
