+++ src/routes/ProductsList/index.jsx
@@ -1,7 +1,7 @@
 import { useLocation } from "react-router-dom";
 import { Helmet, HelmetProvider } from "react-helmet-async";
 import styles from "./ProductsList.module.css";
-import ProductsToDisplay from "../../components/ProductCards/ProductLists";
+import ProductsToDisplay from "../../components/Products/ProductLists";
 import caseFirstLetter from "../../caseFirstLetter";
 
 export default function ProductsList() {
